Changed FrameTime's type to Uint32

+++ OpenGLTesting_C/macOS/Timing.h
@@ -4,7 +4,7 @@ typedef struct FpsLimiter {
     // Variables
     float _fps;
     float _maxFPS;
-    float _frameTime;
+    Uint32 _frameTime;
     Uint32 _startTicks;
 } FpsLimiter;
